+++ src/application.py
@@ -1,8 +1,8 @@
-from telegram.ext import Application, CommandHandler, filters, MessageHandler
+from telegram.ext import Application, CommandHandler
 
 from bot.conversations.command_application import help, stop
 from bot.core.settings import settings
-from bot.handlers import greeting, reg_form_callback
+from bot.handlers import greeting, reg_form_handler
 
 
 def main() -> None:
@@ -11,12 +11,7 @@ def main() -> None:
     app.add_handler(CommandHandler('start', greeting))
     app.add_handler(CommandHandler('help', help))
     app.add_handler(CommandHandler("stop", stop))
-    app.add_handler(
-        MessageHandler(
-            filters.Text(['Регламенты и формы']),
-            reg_form_callback
-        )
-    )
+    app.add_handler(reg_form_handler)
 
     app.run_polling()
